Fix mistake in homopolymer analysis
Fix mistake in identify_indels_within_and_outside_homopolimer_repeats.R
script. This changed the results in Fig. 2C, Clair3-mix, deletion
calls, to be like it used to be before running all scripts at the
second time, which is a more consistent result. The conclusion in the
paper for this new result is: 'Unsurprisingly, the performance of all
pipelines dropped as the length of homopolymer increased'.